requirements.txt
# jupyterコマンドを使う
notebook

numpy
torch
torchvision
scikit-image        # skimage
pillow              # PIL
matplotlib
albumentations
tqdm
seaborn